login register problem fixed , auth guard implemented, add to cart functionality implemented

+++ src/app/dashboard/admin-dashboard/admin-dashboard.component.ts
@@ -2,6 +2,7 @@ import { Input } from '@angular/core';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { AdminService } from 'src/app/products/service/admin.service';
+import { MessageService } from 'src/app/products/service/message.service';
 import { dataModel } from './admin.model';
 
 @Component({
@@ -12,7 +13,8 @@ import { dataModel } from './admin.model';
 export class AdminDashboardComponent implements OnInit {
   dataModelObj : dataModel = new dataModel()
   userData !: any;
-  constructor(private api: AdminService) { }
+  
+  constructor(private api: AdminService, private msgService: MessageService) { }
 
   ngOnInit(): void {
     this.getUserData();
@@ -23,15 +25,19 @@ export class AdminDashboardComponent implements OnInit {
         console.log(res);
       });
     }
-    deleteUserData(data : any){
-      this.api.deleteData(data.id)
-      .subscribe(res=>{
-        alert("Employee Deleted");
-        this.getUserData();
+          deleteUserData(data : any){
+        let conf = confirm(`${data.username}'s data will be deleted `)
+        if (conf){
+        this.api.deleteData(data.id)
+        .subscribe(res=>{
+          this.msgService.showSuccessMessage("User data deleted !")
+            this.api.getData().subscribe(res=>{
+              this.userData=res
+          this.getUserData();
+        })
       })
+      }
     }
-    
-  }
-  
+}
